Extract presenter logic to BaseTradeStateMainChain3bPresenter to avoid code duplication
